requirements.txt
Django==1.11.15
django-nose==1.4.4
django-autoslug==1.9.3
django-taggit==0.21.3
loremipsum==1.0.5
elasticsearch==2.4
django-haystack==2.7.0
-e git+https://github.com/lfalvarez/django-candidator.git#egg=django-candidator
pyyaml>=4.2b1
django-markdown-deux==1.0.4
django-extensions==1.7.5
sorl-thumbnail==12.4a1
Pillow==4.1.1
django-tinymce==2.7.0
django-registration==2.0.4
django-registration-defaults==0.4.4
mimeparse==0.1.3
django-tastypie==0.13.3
linaro-django-pagination==2.0.3
celery==3.1.18
python-memcached==1.58
-e git+https://github.com/mysociety/django-popolo.git#egg=django-popolo
django-mathfilters==0.4.0
django-picklefield==0.3.2
-e git+https://github.com/dokterbob/django-newsletter#egg=django_newsletter
djangorestframework==3.6.2
slumber==0.7.1
-e git+https://github.com/liberation/django-ogp.git#egg=django_ogp
mock==1.3.0
email_reply_parser==0.3.0
flufl.bounce==2.3
django-choices==1.4.2
libsass==0.11.0
django-compressor==2.0
django-sass-processor==0.3.4
facebook-sdk==1.0.0
vcrpy==1.7.4
-e git+https://github.com/mysociety/django-images#egg=django-images
django-bootstrap3==8.2.2
django-formtools==2.0
django-filter==1.0.2
django-el-pagination==3.2.3
django-debug-toolbar==1.6
django-debug-panel==0.8.3
pyjwkest==1.3.2
social-auth-app-django==1.2.0
django-constance==2.0.0
redis==2.10.5
django-redis==4.9.0
pybars3==0.9.3
pytimeparse==1.1.6
django-test-without-migrations==0.6
django-model-utils==3.0.0
django-multiselectfield==0.1.7
django-robots==3.0
djangorestframework-csv==2.0.0
numpy==1.13.3
roman==2.0.0
django-bootstrap4==0.0.6
requests==2.20.1
Collectfast==0.6.2
django-hitcount==1.3.0
django-braces==1.13.0
urllib3>=1.23
Whoosh==2.7.4
